PTSD-523. Add metrics aggregation.

+++ src/basecaller/dataTypes/HQRFPhysicalStates.h
@@ -34,7 +34,7 @@ namespace PacBio {
 namespace Mongo {
 namespace Data {
 
-enum class HQRFPhysicalStates : std::uint16_t
+enum class HQRFPhysicalStates : std::uint8_t
 {
     EMPTY = 0,
     SINGLE,
